requirements.txt
PyQt6>=6.6.0
PyQt6-sip>=13.6.0
pymupdf>=1.23.0
Pillow>=10.0.0
pytesseract>=0.3.10
